Moved dependencies declaration from plugin to simple class Updated example in comments
