+++ buildSrc/src/main/groovy/org/apache/beam/gradle/BeamModulePlugin.groovy
@@ -45,12 +45,12 @@ import org.gradle.testing.jacoco.tasks.JacocoReport
 
 import java.util.concurrent.atomic.AtomicInteger
 
-import static org.apache.beam.gradle.DependenciesPlugin.autovalue_version
-import static org.apache.beam.gradle.DependenciesPlugin.checkerframework_version
-import static org.apache.beam.gradle.DependenciesPlugin.grpc_version
-import static org.apache.beam.gradle.DependenciesPlugin.jsr305_version
-import static org.apache.beam.gradle.DependenciesPlugin.protobuf_version
-import static org.apache.beam.gradle.DependenciesPlugin.spotbugs_version
+import static org.apache.beam.gradle.Dependencies.autovalue_version
+import static org.apache.beam.gradle.Dependencies.checkerframework_version
+import static org.apache.beam.gradle.Dependencies.grpc_version
+import static org.apache.beam.gradle.Dependencies.jsr305_version
+import static org.apache.beam.gradle.Dependencies.protobuf_version
+import static org.apache.beam.gradle.Dependencies.spotbugs_version
 
 /**
  * This plugin adds methods to configure a module with Beam's defaults, called "natures".
